+++ changelog.md
@@ -2,7 +2,7 @@
 
 ### Version 0.2.0 (Beta-3) - TBA
 - New Feature: Introduced paging
-- Improvement: New client API
+- Improvement: Simplified network protocol by removing the duplicate success and response packages
 
 ### Version 0.2.0 (Beta-2) - 28.11.2016
 - New Feature: Java 8 is now required to build the project